requirements.txt
requests
pandas
geopy
python-dotenv
matplotlib
folium
